Changed and fixed the task views is_owner property to and is_editor.
The is_editor function determines whether a given user can edit a task depends on the state of the task. While in review the person who wrote the task, the `owner`, can edit the task, but after the review is complete only an admin, manager, or if it is a subtask to a solution, the owner of that solution may edit the task. This was not being handled by previous solution.